fixed mistakes

+++ spring-10/src/main/java/ru/otus/hw/services/CommentServiceImpl.java
@@ -11,7 +11,6 @@ import ru.otus.hw.repositories.BookRepository;
 import ru.otus.hw.repositories.CommentRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -25,14 +24,15 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<CommentDto> findById(long id) {
-        return commentRepository.findById(id).map(commentConverter::fromDomainObject);
+    public CommentDto findById(long id) {
+        return commentConverter.fromDomainObject(getComment(id));
     }
 
     @Override
     @Transactional(readOnly = true)
     public List<CommentDto> findAllByBookId(long bookId) {
-        return commentRepository.findAllByBookId(bookId).stream()
+        return commentRepository.findAllByBookId(bookId)
+                .stream()
                 .map(commentConverter::fromDomainObject)
                 .toList();
     }
@@ -52,7 +52,15 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public void deleteById(long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Comment with id %d not found".formatted(id));
+        }
         commentRepository.deleteById(id);
+    }
+
+    private Comment getComment(long id) {
+        return commentRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Comment with id %d not found".formatted(id)));
     }
 
     private Comment save(long id, String message, long bookId) {
